+++ FlockRN/services/prayer/prayerService.ts
@@ -18,6 +18,7 @@ import {
   QueryDocumentSnapshot,
   DocumentData,
   CollectionReference,
+  Firestore,
 } from 'firebase/firestore';
 import { db } from '../../firebase/firebaseConfig';
 import {
@@ -26,27 +27,32 @@ import {
   CreatePrayerDTO,
   UpdatePrayerDTO,
   PrayerTopic,
-  CreatePrayerTopicDTO,
-  UpdatePrayerTopicDTO,
   PartialLinkedPrayerEntity,
 } from '@/types/firebase';
-import { PrayerType, EntityType } from '@/types/PrayerSubtypes';
+import { EntityType } from '@/types/PrayerSubtypes';
 import { FirestoreCollections } from '@/schema/firebaseCollections';
 import { getFunctions, httpsCallable } from 'firebase/functions';
 import { getApp } from 'firebase/app';
 import { getEntityType } from '@/types/typeGuards';
 
-class PrayerService {
-  private prayersCollection = collection(db, FirestoreCollections.PRAYERS);
-  private prayerPointsCollection = collection(
-    db,
-    FirestoreCollections.PRAYERPOINTS,
-  );
-  private feedsCollection = collection(db, FirestoreCollections.FEED);
-  private prayerTopicsCollection = collection(
-    db,
-    FirestoreCollections.PRAYERTOPICS,
-  );
+export interface IPrayerService {
+  createPrayer(data: CreatePrayerDTO): Promise<string>;
+  getPrayer(prayerId: string): Promise<Prayer | null>;
+  getUserPrayers(userId: string): Promise<Prayer[]>;
+  updatePrayer(prayerId: string, data: UpdatePrayerDTO): Promise<void>;
+  deletePrayer(prayerId: string, authorId: string): Promise<void>;
+  findRelatedPrayers(
+    embedding: number[],
+    userId: string,
+    sourcePrayerId?: string,
+  ): Promise<PartialLinkedPrayerEntity[] | []>;
+}
+class PrayerService implements IPrayerService {
+  private prayersCollection: CollectionReference<DocumentData>;
+  constructor(db: Firestore) {
+    // Initialize the prayers collection reference
+    this.prayersCollection = collection(db, FirestoreCollections.PRAYERS);
+  }
 
   // ===== Prayer CRUD operations =====
   async createPrayer(data: CreatePrayerDTO): Promise<string> {
@@ -178,111 +184,6 @@ class PrayerService {
     }
   }
 
-  // ==== Prayer Topic CRUD ====
-  // Sharing is not supported for prayer topics yet.
-
-  async createPrayerTopic(data: CreatePrayerTopicDTO): Promise<string> {
-    try {
-      const now = Timestamp.now();
-
-      const newDocRef = doc(this.prayerTopicsCollection);
-      await setDoc(newDocRef, {
-        ...data,
-        id: newDocRef.id,
-        createdAt: now,
-        updatedAt: now,
-        entityType: EntityType.PrayerTopic,
-      });
-      return newDocRef.id;
-    } catch (error) {
-      console.error('Error creating prayer topic:', error);
-      throw error;
-    }
-  }
-
-  //added for PrayerPoint CRUD
-  async updatePrayerTopic(
-    prayerTopicId: string,
-    data: Partial<UpdatePrayerTopicDTO>,
-  ): Promise<void> {
-    try {
-      const now = Timestamp.now();
-      const prayerTopicRef = doc(this.prayerTopicsCollection, prayerTopicId);
-
-      // Check if the prayer topic exists
-      const exists = await this.checkIfDocumentExists(
-        this.prayerTopicsCollection,
-        prayerTopicId,
-      );
-      if (!exists) {
-        throw new Error('Prayer topic not found');
-      }
-
-      await updateDoc(prayerTopicRef, {
-        ...data,
-        updatedAt: now,
-      });
-
-      // Placeholder to add to another feed later.
-    } catch (error) {
-      console.error('Error updating prayer point:', error);
-      throw error;
-    }
-  }
-
-  async deletePrayerTopic(prayerTopicId: string): Promise<void> {
-    try {
-      // Check if the prayer topic exists
-      const exists = await this.checkIfDocumentExists(
-        this.prayerTopicsCollection,
-        prayerTopicId,
-      );
-      if (!exists) {
-        throw new Error('Prayer topic not found');
-      }
-
-      // Delete the prayer topic document
-      await deleteDoc(doc(this.prayerTopicsCollection, prayerTopicId));
-
-      // Placeholder to remove from author's feed later.
-    } catch (error) {
-      console.error('Error deleting prayer topic:', error);
-      throw error;
-    }
-  }
-
-  async getUserPrayerTopics(userId: string): Promise<PrayerTopic[]> {
-    try {
-      const q = query(
-        this.prayerTopicsCollection,
-        where('authorId', '==', userId),
-        orderBy('createdAt', 'desc'),
-      );
-      const querySnapshot = await getDocs(q);
-      if (querySnapshot.empty) return [];
-      return querySnapshot.docs.map(
-        (doc) => this.convertDocToPrayerTopic(doc) as PrayerTopic,
-      );
-    } catch (error) {
-      console.error('Error getting user prayer topics:', error);
-      throw error;
-    }
-  }
-
-  async getPrayerTopic(id: string): Promise<PrayerTopic | null> {
-    try {
-      const docRef = doc(this.prayerTopicsCollection, id);
-      const docSnap = await getDoc(docRef);
-
-      if (!docSnap.exists()) return null;
-
-      return this.convertDocToPrayerTopic(docSnap);
-    } catch (error) {
-      console.error('Error getting prayer points:', error);
-      throw error;
-    }
-  }
-
   // ==== Searchable prayer =====
 
   // Search suggested prayer points
@@ -375,30 +276,6 @@ class PrayerService {
       entityType: data.entityType as EntityType,
     };
   }
-
-  private convertDocToPrayerTopic(
-    docSnap: QueryDocumentSnapshot<DocumentData, DocumentData>,
-  ): PrayerTopic {
-    const data = docSnap.data();
-    return {
-      id: docSnap.id,
-      authorId: data.authorId,
-      authorName: data.authorName,
-      title: data.title,
-      createdAt: data.createdAt as Date,
-      updatedAt: data.updatedAt as Date,
-      prayerTypes: data.prayerTypes as PrayerType[],
-      status: data.status,
-      privacy: data.privacy,
-      recipientId: data.recipientId,
-      recipientName: data.recipientName,
-      journey: data.journey,
-      contextAsStrings: data.contextAsStrings,
-      contextAsEmbeddings: data.contextAsEmbeddings,
-      entityType: data.entityType as EntityType,
-      content: data.content,
-    };
-  }
 }
 
-export const prayerService = new PrayerService();
+export const prayerService = new PrayerService(db);
